fix(messageq): add CI matrix entry and resolve lint and provider test failures.
