+++ .github/workflows/acceptance-tests.yaml
@@ -47,6 +47,7 @@ jobs:
           - keymanager
           - lb
           - marketplace
+          - messageq
           - mnq
           - mongodb
           - object
@@ -217,6 +218,7 @@ jobs:
           - keymanager
           - lb
           - marketplace
+          - messageq
           - mnq
           - mongodb
           - object
