+++ .github/workflows/nightly.yml
@@ -44,6 +44,7 @@ jobs:
           - keymanager
           - lb
           - marketplace
+          - messageq
           - mnq
           - mongodb
           - object
