+++ internal/services/messageq/deployment_data_source_test.go
@@ -98,94 +98,3 @@ data "scaleway_messageq_deployment" "by_id" {
 		},
 	})
 }
-
-func TestAccDataSourceMessageQCertificateAuthority_Basic(t *testing.T) {
-	tt := acctest.NewTestTools(t)
-	defer tt.Cleanup()
-
-	latestVersion := fetchLatestVersion(tt)
-	nodeType := fetchAvailableNodeType(tt)
-
-	resource.ParallelTest(t, resource.TestCase{
-		ProtoV6ProviderFactories: tt.ProviderFactories,
-		CheckDestroy:             isDeploymentDestroyed(tt),
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(`
-resource "scaleway_messageq_deployment" "main" {
-  name       = "tf-test-ds-messageq-ca"
-  version    = "%s"
-  node_count = 1
-  node_type  = "%s"
-  user_name  = "%s"
-  password   = "ThisIsASecurePassword123!"
-  volume {
-    type       = "sbs_5k"
-    size_in_gb = 5
-  }
-}
-
-data "scaleway_messageq_certificate_authority" "main" {
-  deployment_id = scaleway_messageq_deployment.main.id
-}
-`, latestVersion, nodeType, deploymentTestUserName),
-				Check: resource.ComposeTestCheckFunc(
-					isDeploymentPresent(tt, "scaleway_messageq_deployment.main"),
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_certificate_authority.main", "pem"),
-					resource.TestCheckResourceAttrPair(
-						"data.scaleway_messageq_certificate_authority.main", "deployment_id",
-						"scaleway_messageq_deployment.main", "id",
-					),
-				),
-			},
-		},
-	})
-}
-
-func TestAccDataSourceMessageQNodeType_Basic(t *testing.T) {
-	tt := acctest.NewTestTools(t)
-	defer tt.Cleanup()
-
-	nodeType := fetchAvailableNodeType(tt)
-
-	resource.ParallelTest(t, resource.TestCase{
-		ProtoV6ProviderFactories: tt.ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(`
-data "scaleway_messageq_node_type" "main" {
-  name = "%s"
-}
-`, nodeType),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("data.scaleway_messageq_node_type.main", "name", nodeType),
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_node_type.main", "vcpus"),
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_node_type.main", "memory_size_in_gb"),
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_node_type.main", "stock_status"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccDataSourceMessageQVersion_Latest(t *testing.T) {
-	tt := acctest.NewTestTools(t)
-	defer tt.Cleanup()
-
-	resource.ParallelTest(t, resource.TestCase{
-		ProtoV6ProviderFactories: tt.ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: `
-data "scaleway_messageq_version" "latest" {
-  name = "latest"
-}
-`,
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_version.latest", "version"),
-					resource.TestCheckResourceAttrSet("data.scaleway_messageq_version.latest", "id"),
-				),
-			},
-		},
-	})
-}
