+++ internal/services/messageq/version_data_source_test.go
@@ -1,0 +1,30 @@
+package messageq_test
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/scaleway/terraform-provider-scaleway/v2/internal/acctest"
+)
+
+func TestAccDataSourceMessageQVersion_Latest(t *testing.T) {
+	tt := acctest.NewTestTools(t)
+	defer tt.Cleanup()
+
+	resource.ParallelTest(t, resource.TestCase{
+		ProtoV6ProviderFactories: tt.ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: `
+data "scaleway_messageq_version" "latest" {
+  name = "latest"
+}
+`,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrSet("data.scaleway_messageq_version.latest", "version"),
+					resource.TestCheckResourceAttrSet("data.scaleway_messageq_version.latest", "id"),
+				),
+			},
+		},
+	})
+}
